# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KittenClaws",
    "version": "1.0.0",
    "summary": "Kitties got claws...beware",
    "description": "Every response, including errors, is JSON. Errors are `{\"errorMessage\": \"...\"}`: 400 for a malformed or non-object body, a missing, empty or non-string `name`, any other field, or an invalid `X-User-Id` header; 404 for an unknown path or an id that is unknown, soft-deleted or not a UUID; 405 for a method a path does not list (when the request reaches the function); 500, with the details logged, for anything else. `X-User-Id` is taken as sent and is not authenticated.",
    "contact": {
      "name": "Sir Meowsalots"
    },
    "license": {
      "name": "MIT License",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://{restApiId}.execute-api.{region}.amazonaws.com/Prod",
      "description": "Deployed API Gateway stage; its routes have no prefix.",
      "variables": {
        "restApiId": {
          "default": "my-api-id",
          "description": "REST API id from the stack's API URL output."
        },
        "region": {
          "default": "us-east-1"
        }
      }
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local `sam local start-api`, which does not check API keys."
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "tags": [
    {
      "name": "Service",
      "description": "Operational endpoints."
    },
    {
      "name": "Cat",
      "description": "Cat records, stored in the `animals` container."
    },
    {
      "name": "Dog",
      "description": "Dog records, stored in the `animals` container."
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Service"
        ],
        "operationId": "getHealth",
        "summary": "Health check",
        "security": [],
        "responses": {
          "200": {
            "description": "The service is up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/cats": {
      "get": {
        "tags": [
          "Cat"
        ],
        "operationId": "listCat",
        "summary": "List Cat records",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "The records that are not deleted (`[]` when there are none).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cat"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Cat"
        ],
        "operationId": "createCat",
        "summary": "Create a Cat",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cat"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/cats/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Id"
        }
      ],
      "get": {
        "tags": [
          "Cat"
        ],
        "operationId": "getCatById",
        "summary": "Get a Cat",
        "responses": {
          "200": {
            "description": "The record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cat"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "tags": [
          "Cat"
        ],
        "operationId": "updateCat",
        "summary": "Update a Cat's fields",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cat"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Cat"
        ],
        "operationId": "deleteCat",
        "summary": "Soft-delete a Cat",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "The record is marked deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/dogs": {
      "get": {
        "tags": [
          "Dog"
        ],
        "operationId": "listDog",
        "summary": "List Dog records",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "The records that are not deleted (`[]` when there are none).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dog"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Dog"
        ],
        "operationId": "createDog",
        "summary": "Create a Dog",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/dogs/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/Id"
        }
      ],
      "get": {
        "tags": [
          "Dog"
        ],
        "operationId": "getDogById",
        "summary": "Get a Dog",
        "responses": {
          "200": {
            "description": "The record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": [
          "Dog"
        ],
        "operationId": "replaceDog",
        "summary": "Replace a Dog",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogReplaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The replaced record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dog"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Dog"
        ],
        "operationId": "deleteDog",
        "summary": "Soft-delete a Dog",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "The record is marked deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "The API key created by the SAM usage plan."
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Record id; anything other than a UUID is answered with 404.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Most records to return. Values above 1000 are capped at 1000; invalid values fall back to 100.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 100
        }
      },
      "UserId": {
        "name": "X-User-Id",
        "in": "header",
        "required": false,
        "description": "Caller id stored as `createdBy` (on create) and `updatedBy`; surrounding whitespace is trimmed and a write without it removes `updatedBy`. Not authenticated.",
        "schema": {
          "type": "string",
          "maxLength": 256
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The body or the `X-User-Id` header is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "No record has this id, it is deleted, or the id is not a UUID.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "An unexpected error; the details are logged, never returned.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "errorMessage"
        ],
        "properties": {
          "errorMessage": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Health": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          }
        },
        "additionalProperties": false
      },
      "DeleteResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "examples": [
              "<Name> with id <id> was deleted successfully."
            ]
          }
        },
        "additionalProperties": false
      },
      "Cat": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CatCreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "CatUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "Dog": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DogCreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "DogReplaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      }
    }
  }
}